Moving controlled test into testing utils
