+++ ui/components/component-library/text-field-search/text-field-search.test.js
@@ -1,7 +1,7 @@
 /* eslint-disable jest/require-top-level-describe */
 import React from 'react';
-import { fireEvent, render } from '@testing-library/react';
-import setupControlled from '../text-field/text-field.test';
+import { render } from '@testing-library/react';
+import { renderControlledInput } from '../../../../test/lib/render-helpers';
 import { TextFieldSearch } from './text-field-search';
 
 describe('TextFieldSearch', () => {
@@ -10,8 +10,8 @@ describe('TextFieldSearch', () => {
     expect(getByRole('searchbox')).toBeDefined();
   });
   it('should render showClearButton button when showClearButton is true and value exists', async () => {
-    // As showClearButton is intended to be used with a controlled input we need to use setupControlled
-    const { user, getByRole } = setupControlled(TextFieldSearch, {
+    // As showClearButton is intended to be used with a controlled input we need to use renderControlledInput
+    const { user, getByRole } = renderControlledInput(TextFieldSearch, {
       showClearButton: true,
     });
     await user.type(getByRole('searchbox'), 'test value');
@@ -19,20 +19,23 @@ describe('TextFieldSearch', () => {
     expect(getByRole('button', { name: /Clear/u })).toBeDefined();
   });
   it('should still render with the rightAccessory when showClearButton is true', async () => {
-    // As showClearButton is intended to be used with a controlled input we need to use setupControlled
-    const { user, getByRole, getByText } = setupControlled(TextFieldSearch, {
-      showClearButton: true,
-      rightAccessory: <div>right-accessory</div>,
-    });
+    // As showClearButton is intended to be used with a controlled input we need to use renderControlledInput
+    const { user, getByRole, getByText } = renderControlledInput(
+      TextFieldSearch,
+      {
+        showClearButton: true,
+        rightAccessory: <div>right-accessory</div>,
+      },
+    );
     await user.type(getByRole('searchbox'), 'test value');
     expect(getByRole('searchbox')).toHaveValue('test value');
     expect(getByRole('button', { name: /Clear/u })).toBeDefined();
     expect(getByText('right-accessory')).toBeDefined();
   });
   it('should fire onClick event when passed to clearButtonOnClick when clear button is clicked', async () => {
-    // As showClearButton is intended to be used with a controlled input we need to use setupControlled
+    // As showClearButton is intended to be used with a controlled input we need to use renderControlledInput
     const fn = jest.fn();
-    const { user, getByRole } = setupControlled(TextFieldSearch, {
+    const { user, getByRole } = renderControlledInput(TextFieldSearch, {
       showClearButton: true,
       clearButtonOnClick: fn,
     });
@@ -41,9 +44,9 @@ describe('TextFieldSearch', () => {
     expect(fn).toHaveBeenCalledTimes(1);
   });
   it('should fire onClick event when passed to clearButtonProps.onClick prop', async () => {
-    // As showClearButton is intended to be used with a controlled input we need to use setupControlled
+    // As showClearButton is intended to be used with a controlled input we need to use renderControlledInput
     const fn = jest.fn();
-    const { user, getByRole } = setupControlled(TextFieldSearch, {
+    const { user, getByRole } = renderControlledInput(TextFieldSearch, {
       showClearButton: true,
       clearButtonProps: { onClick: fn },
     });
